+++ fixtures/blocks/src/client/Shell.js
@@ -6,54 +6,14 @@
  */
 
 import * as React from 'react';
-import Link from './Link';
-import {useRouter} from './RouterContext';
-
-function TabBar({children}) {
-  return (
-    <div style={{border: '1px solid #aaa', padding: 20, width: 500}}>
-      {children}
-    </div>
-  );
-}
-
-function TabLink({to, children}) {
-  const {pendingUrl: activeUrl} = useRouter();
-  const active = activeUrl === to;
-  if (active) {
-    return (
-      <b
-        style={{
-          display: 'inline-block',
-          width: 50,
-          marginRight: 20,
-        }}>
-        {children}
-      </b>
-    );
-  }
-  return (
-    <Link
-      style={{
-        display: 'inline-block',
-        width: 50,
-        marginRight: 20,
-      }}
-      to={to}>
-      {children}
-    </Link>
-  );
-}
+import TabNav from './TabNav';
 
 // TODO: Error Boundaries.
 
 export default function Shell({children}) {
   return (
     <>
-      <TabBar>
-        <TabLink to="/">Home</TabLink>
-        <TabLink to="/profile/3">Profile</TabLink>
-      </TabBar>
+      <TabNav />
       <br />
       {children}
     </>
